refactor: extract ErgTooltip, StrengthTooltip, LoadTooltip components
Moves three chart tooltip components (~134 lines total) out of the monolith.
fmtPace helper moves with ErgTooltip (its sole caller).

Build and 68 tests pass.
